+++ polar-core/src/polar.rs
@@ -1,4 +1,5 @@
 use super::error::PolarResult;
+use super::messages::*;
 use super::parser;
 use super::rewrites::*;
 use super::types::*;
